# packages/contracts/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoCoLo API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/events": {
      "get": {
        "operationId": "listEvents",
        "summary": "期間内の予定一覧を取得",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "予定一覧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "operationId": "createEvent",
        "summary": "予定を登録",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "予定を登録",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventMutationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/events/{eventId}": {
      "get": {
        "operationId": "getEvent",
        "summary": "予定詳細を取得",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "予定詳細",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventMutationResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "patch": {
        "operationId": "updateEvent",
        "summary": "予定を編集",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "予定を更新",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventMutationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/events/{eventId}/attendance": {
      "get": {
        "operationId": "getCurrentAttendance",
        "summary": "現在の出欠回答を取得",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "現在の出欠回答",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "put": {
        "operationId": "upsertEventAttendance",
        "summary": "予定の出欠を登録または修正",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "出欠回答",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/events/{eventId}/attendance/summary": {
      "get": {
        "operationId": "getEventAttendanceSummary",
        "summary": "予定の出欠集計を取得",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "出欠集計",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSummaryResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "EventInput": {
        "type": "object",
        "required": [
          "title",
          "type",
          "startsAt",
          "endsAt",
          "fee",
          "transportationRequired",
          "attendanceDeadline"
        ],
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "type": {
            "type": "string",
            "enum": [
              "practice",
              "match",
              "event"
            ]
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "itemsToBring": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "fee": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "announcementImageAttachmentId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "opponent": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          },
          "meetingTime": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "transportationRequired": {
            "type": "boolean"
          },
          "attendanceDeadline": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EventUpdateInput": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "type": {
            "type": "string",
            "enum": [
              "practice",
              "match",
              "event"
            ]
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "itemsToBring": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "fee": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "announcementImageAttachmentId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "opponent": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          },
          "meetingTime": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "transportationRequired": {
            "type": "boolean"
          },
          "attendanceDeadline": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EventResponse": {
        "type": "object",
        "required": [
          "id",
          "title",
          "type",
          "startsAt",
          "endsAt",
          "location",
          "itemsToBring",
          "fee",
          "announcementImageAttachmentId",
          "opponent",
          "meetingTime",
          "transportationRequired",
          "attendanceDeadline",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "type": {
            "type": "string",
            "enum": [
              "practice",
              "match",
              "event"
            ]
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "itemsToBring": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "fee": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "announcementImageAttachmentId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "opponent": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          },
          "meetingTime": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "transportationRequired": {
            "type": "boolean"
          },
          "attendanceDeadline": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EventListResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "maxItems": 500,
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            }
          }
        }
      },
      "EventMutationResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EventResponse"
          }
        }
      },
      "AttendanceInput": {
        "type": "object",
        "required": [
          "response"
        ],
        "additionalProperties": false,
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "response": {
            "type": "string",
            "enum": [
              "attending",
              "absent",
              "pending"
            ]
          },
          "correctionReason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "oneOf": [
          {
            "required": [
              "memberId"
            ]
          },
          {
            "required": [
              "subjectMemberId"
            ]
          }
        ]
      },
      "AttendanceResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "eventId",
              "memberId",
              "response",
              "updatedAt"
            ],
            "additionalProperties": false,
            "properties": {
              "eventId": {
                "type": "string",
                "format": "uuid"
              },
              "memberId": {
                "type": "string",
                "format": "uuid"
              },
              "response": {
                "type": "string",
                "enum": [
                  "attending",
                  "absent",
                  "pending"
                ]
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "AttendanceListResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "eventId",
                "memberId",
                "response",
                "updatedAt"
              ],
              "additionalProperties": false,
              "properties": {
                "eventId": {
                  "type": "string",
                  "format": "uuid"
                },
                "memberId": {
                  "type": "string",
                  "format": "uuid"
                },
                "response": {
                  "type": "string",
                  "enum": [
                    "attending",
                    "absent",
                    "pending"
                  ]
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
      "AttendanceSummaryResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "totalMembers",
              "attending",
              "absent",
              "pending",
              "unanswered",
              "unansweredMemberIds"
            ],
            "additionalProperties": false,
            "properties": {
              "totalMembers": {
                "type": "integer",
                "minimum": 0
              },
              "attending": {
                "type": "integer",
                "minimum": 0
              },
              "absent": {
                "type": "integer",
                "minimum": 0
              },
              "pending": {
                "type": "integer",
                "minimum": 0
              },
              "unanswered": {
                "type": "integer",
                "minimum": 0
              },
              "unansweredMemberIds": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "requestId"
            ],
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              "details": {},
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "入力値が不正です。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "認証が必要です。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "この操作を実行する権限がありません。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "リクエストが現在の状態と競合しました。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "指定されたリソースが見つかりません。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "リクエスト数の上限を超えました。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "内部エラーが発生しました。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "外部サービスまたは依存サービスが利用できません。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
...
}